refactorize workspace funcionality
