+++ packages/browser/src/Component/Widget/Server/Server.js
@@ -1,0 +1,34 @@
+// API @ud-viz/browser/src/Component/Itowns/Widget/Server
+
+// Collection of Itowns Widget using an external server
+
+export { AuthenticationService } from './Authentication/services/AuthenticationService';
+
+export { AuthenticationView } from './Authentication/views/AuthenticationView';
+
+export { DocumentCommentsModule } from './DocumentComments/DocumentCommentsModule';
+
+export { GeocodingService } from './Geocoding/services/GeocodingService';
+
+export { GeocodingView } from './Geocoding/views/GeocodingView';
+
+export { ContributeModule } from './Contribute/ContributeModule';
+
+export { DocumentValidationModule } from './DocumentValidation/DocumentValidationModule';
+
+export { DocumentVisualizerWindow } from './DocumentVisualizer/View/DocumentVisualizerWindow';
+
+export { SparqlWidgetView } from './SPARQL/View/SparqlWidgetView';
+
+export { SparqlEndpointResponseProvider } from './SPARQL/Service/SparqlEndpointResponseProvider';
+
+export { SparqlWorkspaceWidgetView } from './Workspace/View/SparqlWorkspaceWidgetView';
+
+export { GuidedTourController } from './GuidedTour/GuidedTourController';
+
+export { LinkModule } from './Links/LinkModule';
+
+export { DocumentModule } from './Documents/DocumentModule';
+
+import * as Component from './Component/RequestService';
+export { Component };
